requirements.txt
boto3
botocore
catppuccin
certifi
colorlog
dill
python-dotenv
fastapi
from_root
halo
imblearn
ipykernel
jinja2
matplotlib
mypy-boto3-s3
numpy
pandas
PyYAML
python-multipart
pymongo
scikit-learn
seaborn
setuptools
uvicorn
wheel
